+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Enhancements
 
+- remove redundant error toast([#515](https://github.com/opensearch-project/dashboards-assistant/pull/515))
+- Add auto suggested aggregation for text2Viz ([#514](https://github.com/opensearch-project/dashboards-assistant/pull/514))
+
+
 ### Bug Fixes
 
 ### Infrastructure